# components/org.wso2.carbon.stream.processor.status.dashboard/src/test/resources/swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample Worker server."
  version: "1.0.0"
  title: "Worker "

tags:
- name: "worker"
  description: "Add/delete/update workers"
schemes:
- "http"
paths:
  /worker:
    post:
      tags:
      - "worker"
      summary: "Add a new worker"
      operationId: "addWorker"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "worker"
        description: "Worker object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/Worker"
      responses:
        200:
          description: "Successfully added"
          schema:
            $ref: '#/definitions/Worker'
        405:
          description: "Invalid input"
          
  /worker/{id}:
    get:
      tags:
      - "worker"
      summary: "Find worker by ID"
      description: "Returns a single worker"
      operationId: "getWorkerById"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of worker to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Worker"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Worker not found"
          
    put:
      tags:
      - "worker"
      summary: "Update an existing worker"
      operationId: "updateWorker"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of worker that needs to be updated"
        required: true
        type: "integer"
        format: "int64"
        
      - in: "body"
        name: "worker"
        description: "Worker object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/Worker"
      responses:
        200:
          description: "Successfully updated the worker"
          schema:
            $ref: '#/definitions/Worker'
        
        400:
          description: "Invalid ID supplied"
        404:
          description: "Worker not found"
        405:
          description: "Validation exception"
          
    delete:
      tags:
      - "worker"
      summary: "Deletes a worker"
      operationId: "deleteWorker"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Worker id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Successfully deleted"
        
        400:
          description: "Invalid ID supplied"
        404:
          description: "Worker not found"
  
definitions:
  Worker:
    type: "object"
    required:
    - "username"
    - "password"
    - "url"
    properties:
      id:
        type: "integer"
        format: "int64"
      username:
        type: "string"
      password:
        type: "string"
      url:
        type: "string"
  